requirements.txt
pypng
pyqrcode
python-barcode